Add timeit and traceit decorators for optimization purposes
