+++ pyproject.toml
@@ -16,6 +16,7 @@ black = "^22.10.0"
 isort = "^5.10.1"
 marshmallow-dataclass = "^8.5.9"
 cairo-coverage = "^0.1.4"
+pandas = "^1.5.1"
 
 [tool.pytest.ini_options]
 filterwarnings = [
