+++ tests/conftest.py
@@ -1,5 +1,4 @@
 import asyncio
-from time import time
 from typing import AsyncGenerator
 
 import pytest
@@ -7,6 +6,8 @@ import pytest_asyncio
 from cairo_coverage import cairo_coverage
 from starkware.starknet.business_logic.state.state_api_objects import BlockInfo
 from starkware.starknet.testing.starknet import Starknet
+
+from tests.utils.utils import reports, timeit
 
 
 @pytest.fixture(scope="session")
@@ -22,6 +23,9 @@ async def starknet() -> AsyncGenerator[Starknet, None]:
     starknet.state.state.update_block_info(
         BlockInfo.create_for_testing(block_number=1, block_timestamp=1)
     )
+    starknet.deploy = timeit(starknet.deploy)
+    starknet.declare = timeit(starknet.declare)
+
     yield starknet
     files = cairo_coverage.report_runs(excluded_file={"site-packages", "cairo_files"})
     total_covered = []
@@ -31,6 +35,9 @@ async def starknet() -> AsyncGenerator[Starknet, None]:
         total_covered.append(file.pct_covered)
     if files and (val := not sum(total_covered) / len(files)) >= 80:
         print(f"WARNING: Project is not covered enough {val:.2f})")
+    times, resources = reports()
+    print(times)
+    print(resources)
 
 
 @pytest_asyncio.fixture(scope="session")
@@ -44,17 +51,12 @@ async def eth(starknet):
 
 @pytest_asyncio.fixture(scope="session")
 async def account_registry(starknet: Starknet):
-    print("account_registry")
-    start = time()
-    registry = await starknet.deploy(
+    return await starknet.deploy(
         source="./src/kakarot/accounts/registry/account_registry.cairo",
         cairo_path=["src"],
         disable_hint_validation=False,
         constructor_calldata=[1],
     )
-    registry_time = time()
-    print(f"AccountRegistry deployed in {registry_time - start:.2f}s")
-    return registry
 
 
 @pytest_asyncio.fixture(scope="session")
