+++ tests/integrations/test_cases/erc20.py
@@ -1,45 +1,4 @@
-from textwrap import wrap
-from time import time
-
 import pytest
-import pytest_asyncio
-from starkware.starknet.testing.contract import DeclaredClass, StarknetContract
-from starkware.starknet.testing.starknet import Starknet
-
-
-@pytest_asyncio.fixture(scope="module")
-async def zk_evm(
-    starknet: Starknet, eth: StarknetContract, contract_account_class: DeclaredClass
-):
-    start = time()
-    _zk_evm = await starknet.deploy(
-        source="./src/kakarot/kakarot.cairo",
-        cairo_path=["src"],
-        disable_hint_validation=False,
-        constructor_calldata=[
-            1,
-            eth.contract_address,
-            contract_account_class.class_hash,
-        ],
-    )
-    evm_time = time()
-    print(f"\nzkEVM deployed in {evm_time - start:.2f}s")
-    return _zk_evm
-
-
-@pytest_asyncio.fixture(scope="module", autouse=True)
-async def set_account_registry(zk_evm, account_registry):
-    await account_registry.transfer_ownership(zk_evm.contract_address).execute(
-        caller_address=1
-    )
-    await zk_evm.set_account_registry(
-        registry_address_=account_registry.contract_address
-    ).execute(caller_address=1)
-    yield
-    await account_registry.transfer_ownership(1).execute(
-        caller_address=zk_evm.contract_address
-    )
-
 
 test_cases = [
     {
@@ -59,124 +18,6 @@ test_cases = [
             "return_value": "0000000000000000000000000000000000000000000000000000000000000005",
         },
         "id": "solmate_erc20_all_test",
+        "marks": [pytest.mark.SolmateERC20],
     },
 ]
-
-
-params = [pytest.param(case.pop("params"), **case) for case in test_cases]
-
-
-@pytest.mark.asyncio
-class TestZkEVM:
-    @staticmethod
-    def int_to_uint256(value):
-        low = value & ((1 << 128) - 1)
-        high = value >> 128
-        return low, high
-
-    @pytest.mark.parametrize(
-        "params",
-        params[:1],
-        # TODO: not sure how of those we want to re-run with the execute_at_address because it is very slow
-        # TODO: This is a magic number.
-    )
-    async def test_erc20(
-        self, starknet: Starknet, zk_evm, params, contract_account_class: DeclaredClass
-    ):
-        Uint256 = zk_evm.struct_manager.get_contract_struct("Uint256")
-
-        # Deploy contract by sending calldata to address 0
-        print("DEPLOYING CONTRACT")
-        res = await zk_evm.execute_at_address(
-            address=0,
-            value=0,
-            calldata=[int(b, 16) for b in wrap(params["code"], 2)],
-        ).execute(caller_address=1)
-
-        print("ERC20 Deployment tx infos")
-        print(res.call_info.execution_resources)
-
-        evm_contract_address = res.result.evm_contract_address
-        starknet_contract_address = res.result.starknet_contract_address
-        value = 0
-
-        # Deploy initiate deployed contract
-        print("INITIATING CONTRACT")
-        res = await zk_evm.initiate(
-            evm_address=evm_contract_address,
-            starknet_address=starknet_contract_address,
-            value=value,
-        ).execute(caller_address=1)
-
-        print("ERC20 Contract initialization tx infos")
-        print(res.call_info.execution_resources)
-
-        # Call mint function
-        print("CALLING mint TX")
-        res = await zk_evm.execute_at_address(
-            address=evm_contract_address,
-            value=value,
-            calldata=[int(b, 16) for b in wrap(params["mint"], 2)],
-        ).execute(caller_address=2)
-
-        print("ERC20 mint tx infos")
-        print(res.call_info.execution_resources)
-
-        # Call approve function
-        print("CALLING approve TX")
-        res = await zk_evm.execute_at_address(
-            address=evm_contract_address,
-            value=value,
-            calldata=[int(b, 16) for b in wrap(params["approve"], 2)],
-        ).execute(caller_address=2)
-
-        print("ERC20 approve tx infos")
-        print(res.call_info.execution_resources)
-
-        # Call allowance function
-        print("CALLING allowance TX")
-        res = await zk_evm.execute_at_address(
-            address=evm_contract_address,
-            value=value,
-            calldata=[int(b, 16) for b in wrap(params["allowance"], 2)],
-        ).execute(caller_address=2)
-
-        print("ERC20 allowance tx infos")
-        print(res.call_info.execution_resources)
-
-        # Call transferFrom function
-        print("CALLING transferFrom TX")
-        res = await zk_evm.execute_at_address(
-            address=evm_contract_address,
-            value=value,
-            calldata=[int(b, 16) for b in wrap(params["transferFrom"], 2)],
-        ).execute(caller_address=1)
-
-        print("ERC20 transferFrom tx infos")
-        print(res.call_info.execution_resources)
-
-        # Call transfer function
-        print("CALLING transfer")
-        res = await zk_evm.execute_at_address(
-            address=evm_contract_address,
-            value=value,
-            calldata=[int(b, 16) for b in wrap(params["transfer"], 2)],
-        ).execute(caller_address=1)
-
-        print("ERC20 transfer tx infos")
-        print(res.call_info.execution_resources)
-
-        # Call balanceOf function
-        print("CHECKING balanceOf TX")
-        res = await zk_evm.execute_at_address(
-            address=evm_contract_address,
-            value=value,
-            calldata=[int(b, 16) for b in wrap(params["balanceOf"], 2)],
-        ).execute(caller_address=1)
-
-        print("ERC20 balanceOf tx infos")
-        print(res.call_info.execution_resources)
-
-        assert res.result.return_data == [
-            int(m, 16) for m in wrap(params["return_value"], 2)
-        ]
